backend/requirements.txt
asgiref==3.9.1
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
cryptography==45.0.6
dj-database-url==3.0.1
dj-rest-auth==7.0.1
Django==5.2.5
django-allauth==65.11.0
django-cors-headers==4.7.0
django-stubs==5.2.2
django-stubs-ext==5.2.2
djangorestframework==3.16.1
djangorestframework-stubs==3.16.2
djangorestframework_simplejwt==5.5.1
google-cloud-storage==3.3.0
google-genai>=1.32.0
gunicorn==23.0.0
idna==3.10
mypy==1.17.1
mypy_extensions==1.1.0
oauthlib==3.3.1
packaging==25.0
pathspec==0.12.1
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.10.1
python-dotenv==1.1.1
requests==2.32.4
ruff==0.12.9
sqlparse==0.5.3
types-PyYAML==6.0.12.20250809
types-requests==2.32.4.20250809
typing_extensions==4.14.1
urllib3==2.5.0
django-allow-cidr==0.8.0
python-json-logger==3.3.0
whitenoise==6.9.0

# Celery and task queue dependencies
celery==5.5.3
redis==6.4.0
kombu[gcpubsub]==5.5.4
django-celery-results==2.6.0
